atualizando funcoes para int16 e atualizando mir

+++ arithmetic.h
@@ -2,30 +2,30 @@
 
 using namespace std;
 
-int add(vector<int>& pilha){
+int16_t add(vector<int16_t>& pilha){
     return pilha[pilha.size() - 1] + pilha[pilha.size() - 2];
 }
 
-void inc(vector<int>& pilha){
+void inc(vector<int16_t>& pilha){
     pilha[pilha.size() - 1] += 1;
 }
 
-int sub(vector<int>& pilha){
+int16_t sub(vector<int16_t>& pilha){
     return pilha[pilha.size() - 1] - pilha[pilha.size() - 2];
 }
 
-int mul(vector<int>& pilha){
+int16_t mul(vector<int16_t>& pilha){
     return pilha[pilha.size() - 1] * pilha[pilha.size() - 2];
 }
 
-int divis(vector<int>& pilha){
+int16_t divis(vector<int16_t>& pilha){
     return pilha[pilha.size() - 1] / pilha[pilha.size() - 2];
 }
 
-int mod(vector<int>& pilha){
+int16_t mod(vector<int16_t>& pilha){
     return pilha[pilha.size() - 1] % pilha[pilha.size() - 2];
 }
 
-int sub2(vector<int>& pilha){
+int16_t sub2(vector<int16_t>& pilha){
     return pilha[pilha.size() - 2] - pilha[pilha.size() - 3];
 }
